Problem: non Rust idiomatic error handling
Solution: while passing String is good enough, idiomatic Rust code
prefer passing an enum. It has an advantage that it encodes all errors
states of the code in a single place.

Add ErrorKind enum with CommandNotFound and String variants and
print_error function, which prints the human readable error on console.

Add an integration test for CommandNotFound error.