Added CI tests for dnf5.
This also builds a new docker image with dnf5 installed, but no zypper.
The README is also updated on how to manually build the images and run
the tests.

The bats submodule is also updated to v1.11.0, as that version
introduces a "bats_test_function" to dynamically register test functions.